planner: fix join reorder through projection edge cases
