+++ pkg/planner/core/rule_join_reorder_projection_inline.go
@@ -17,139 +17,14 @@ package core
 import (
 	"github.com/pingcap/tidb/pkg/expression"
 	"github.com/pingcap/tidb/pkg/planner/core/base"
+	"github.com/pingcap/tidb/pkg/planner/core/joinorder"
 	"github.com/pingcap/tidb/pkg/planner/core/operator/logicalop"
 	"github.com/pingcap/tidb/pkg/sessionctx/vardef"
 )
 
 // This file contains helpers for join reorder "look through Projection" optimization:
 // - best-effort inlining checks for Projection on top of Join
-// - derived column mapping (colExprMap) propagation and substitution
-// - shared expression rewrite helpers used by the optimization (and join-key materialization)
-
-// substituteColsInEqEdges substitutes derived columns in equality edges using colExprMap.
-func substituteColsInEqEdges(edges []*expression.ScalarFunction, colExprMap map[int64]expression.Expression) []*expression.ScalarFunction {
-	result := make([]*expression.ScalarFunction, 0, len(edges))
-	for _, edge := range edges {
-		substituted := substituteColsInExpr(edge, colExprMap)
-		if sf, ok := substituted.(*expression.ScalarFunction); ok {
-			result = append(result, sf)
-		} else {
-			result = append(result, edge)
-		}
-	}
-	return result
-}
-
-type exprReplacer func(expr expression.Expression) (newExpr expression.Expression, replaced bool)
-
-// rewriteExprTree rewrites an expression tree in a best-effort, copy-on-write way.
-//
-// The replacer is applied in pre-order (parent before children). If it replaces a node,
-// the returned expression will be rewritten again so callers can implement recursive
-// substitutions (e.g. colExprMap chains) without duplicating traversal logic.
-func rewriteExprTree(expr expression.Expression, replace exprReplacer) expression.Expression {
-	if expr == nil {
-		return nil
-	}
-
-	if replace != nil {
-		if newExpr, replaced := replace(expr); replaced {
-			if newExpr == nil {
-				return nil
-			}
-			if newExpr != expr {
-				return rewriteExprTree(newExpr, replace)
-			}
-		}
-	}
-
-	sf, ok := expr.(*expression.ScalarFunction)
-	if !ok {
-		return expr
-	}
-
-	// Copy-on-write: only clone the function node when any argument changes.
-	oldArgs := sf.GetArgs()
-	var newArgs []expression.Expression
-	for i, arg := range oldArgs {
-		rewrittenArg := rewriteExprTree(arg, replace)
-		if newArgs == nil {
-			if rewrittenArg == arg {
-				continue
-			}
-			newArgs = make([]expression.Expression, len(oldArgs))
-			copy(newArgs, oldArgs[:i])
-		}
-		newArgs[i] = rewrittenArg
-	}
-	if newArgs == nil {
-		return sf
-	}
-
-	newSf := sf.Clone().(*expression.ScalarFunction)
-	args := newSf.GetArgs()
-	for i := range args {
-		args[i] = newArgs[i]
-	}
-	// Args changed: clear cached hash so CanonicalHashCode reflects rewritten children.
-	newSf.CleanHashCode()
-	return newSf
-}
-
-// substituteColsInExprs substitutes derived columns in a list of expressions using colExprMap.
-func substituteColsInExprs(exprs []expression.Expression, colExprMap map[int64]expression.Expression) []expression.Expression {
-	result := make([]expression.Expression, 0, len(exprs))
-	for _, expr := range exprs {
-		result = append(result, substituteColsInExpr(expr, colExprMap))
-	}
-	return result
-}
-
-// substituteExprsWithColsInExprs replaces deterministic sub-expressions in exprs with a
-// pre-materialized column from expr2Col (keyed by expression.CanonicalHashCode()).
-func substituteExprsWithColsInExprs(exprs []expression.Expression, expr2Col map[string]*expression.Column) []expression.Expression {
-	if len(expr2Col) == 0 {
-		return exprs
-	}
-	result := make([]expression.Expression, 0, len(exprs))
-	for _, expr := range exprs {
-		result = append(result, substituteExprsWithColsInExpr(expr, expr2Col))
-	}
-	return result
-}
-
-func substituteExprsWithColsInExpr(expr expression.Expression, expr2Col map[string]*expression.Column) expression.Expression {
-	if len(expr2Col) == 0 {
-		return expr
-	}
-	return rewriteExprTree(expr, func(e expression.Expression) (expression.Expression, bool) {
-		if col, ok := expr2Col[string(e.CanonicalHashCode())]; ok {
-			return col, true
-		}
-		return e, false
-	})
-}
-
-// substituteColsInExpr recursively substitutes derived columns in an expression using colExprMap.
-// It replaces column references with their defining expressions from colExprMap.
-func substituteColsInExpr(expr expression.Expression, colExprMap map[int64]expression.Expression) expression.Expression {
-	if len(colExprMap) == 0 {
-		return expr
-	}
-	return rewriteExprTree(expr, func(e expression.Expression) (expression.Expression, bool) {
-		col, ok := e.(*expression.Column)
-		if !ok {
-			return e, false
-		}
-		if defExpr, ok := colExprMap[col.UniqueID]; ok {
-			// Expressions in colExprMap are treated as immutable in join-reorder flow.
-			// Reuse pointers to avoid extra clones/allocations; if a future pass starts
-			// mutating these expression trees in-place, clone defExpr here before return.
-			return defExpr, true
-		}
-		return e, false
-	})
-}
+// - derived column mapping (colExprMap) propagation and inlining safety checks
 
 // tryInlineProjectionForJoinGroup tries to extract the join group through a Projection on top of a Join.
 //
@@ -197,7 +72,7 @@ func buildColExprMapForProjection(
 		expr := proj.Exprs[i]
 		// Use child's colExprMap to substitute column references in current expression.
 		if len(childColExprMap) > 0 {
-			expr = substituteColsInExpr(expr, childColExprMap)
+			expr = joinorder.SubstituteColsInExpr(expr, childColExprMap)
 		}
 
 		// For pass-through columns (where the projection expression is just a Column reference
@@ -224,56 +99,6 @@ func buildColExprMapForProjection(
 		}
 	}
 	return colExprMap
-}
-
-// outerJoinSideFiltersTouchMultipleLeaves checks whether the outer-join filters depend on more than one
-// leaf on the outer side. If so, we conservatively disable join reordering for this join node.
-//
-// When projections are inlined under the outer side, join conditions may reference derived columns that
-// are not contained in any leaf schema. We substitute those derived columns via `outerColExprMap` before
-// extracting referenced columns.
-func outerJoinSideFiltersTouchMultipleLeaves(
-	join *logicalop.LogicalJoin,
-	outerGroup []base.LogicalPlan,
-	outerColExprMap map[int64]expression.Expression,
-	outerIsLeft bool,
-) bool {
-	if join == nil {
-		return false
-	}
-
-	checkOtherConds := join.OtherConditions
-	checkSideConds := join.RightConditions
-	if outerIsLeft {
-		checkSideConds = join.LeftConditions
-	}
-	checkEQConds := expression.ScalarFuncs2Exprs(join.EqualConditions)
-
-	if len(outerColExprMap) > 0 {
-		checkOtherConds = substituteColsInExprs(checkOtherConds, outerColExprMap)
-		checkSideConds = substituteColsInExprs(checkSideConds, outerColExprMap)
-		checkEQConds = substituteColsInExprs(checkEQConds, outerColExprMap)
-	}
-
-	extractedCols := make(map[int64]*expression.Column, len(checkOtherConds)+len(checkSideConds)+len(checkEQConds))
-	expression.ExtractColumnsMapFromExpressionsWithReusedMap(extractedCols, nil, checkOtherConds...)
-	expression.ExtractColumnsMapFromExpressionsWithReusedMap(extractedCols, nil, checkSideConds...)
-	expression.ExtractColumnsMapFromExpressionsWithReusedMap(extractedCols, nil, checkEQConds...)
-
-	affectedGroups := 0
-	for _, outerLeaf := range outerGroup {
-		leafSchema := outerLeaf.Schema()
-		for _, col := range extractedCols {
-			if leafSchema.Contains(col) {
-				affectedGroups++
-				break
-			}
-		}
-		if affectedGroups > 1 {
-			return true
-		}
-	}
-	return false
 }
 
 // isInlineableProjectionExpr checks whether `expr` only consists of supported expression nodes,
@@ -357,7 +182,13 @@ func canInlineProjection(proj *logicalop.LogicalProjection, childResult *joinGro
 		checkExpr := expr
 		// Inline any previously inlined derived columns first, so we only reason on columns of join-group leaves.
 		if len(childResult.colExprMap) > 0 {
-			checkExpr = substituteColsInExpr(checkExpr, childResult.colExprMap)
+			checkExpr = joinorder.SubstituteColsInExpr(checkExpr, childResult.colExprMap)
+		}
+		if childResult.nullExtendedCols != nil && expression.ExprReferenceSchema(checkExpr, childResult.nullExtendedCols) {
+			// Expressions that reference the null-extended side of an outer join must stay above
+			// that outer join. Inlining them would let later join reorder treat them as normal
+			// single-leaf connectors and can change outer-join semantics.
+			return false
 		}
 
 		cols := expression.ExtractColumns(checkExpr)
